+++ main/CMakeLists.txt
@@ -83,7 +83,7 @@ if (BUILD_UI_MU4)
     set(LINK_LIB ${LINK_LIB}
         actions
         appshell
-        account
+        cloud
         context
         shortcuts
         workspace
