--- mu4/account/account.qrc
@@ -1,6 +1,0 @@
-<RCC>
-    <qresource prefix="/">
-        <file>qml/MuseScore/Account/qmldir</file>
-        <file>qml/MuseScore/Account/AccountInfo.qml</file>
-    </qresource>
-</RCC>
