+++ mu4/cloud/qml/MuseScore/Cloud/qmldir
@@ -1,2 +1,2 @@
-module MuseScore.Account
+module MuseScore.Cloud
 AccountInfo 1.0 AccountInfo.qml
